TFG_Ciberseguridad/bot_python/requirements.txt
# Dependencias del bot de Python
python-telegram-bot>=20.0
pymongo>=4.0
openai>=1.0
watchdog>=3.0
python-dotenv>=1.0
